requirements.txt
python-dotenv>=1.0.1
openai>=1.40.0
pandas>=2.2.2
scikit-learn>=1.5.1
numpy>=1.26.4
tqdm>=4.66.4
joblib>=1.4.2
streamlit>=1.37.0
datasets>=2.20.0
